implement main component level testing
